CTAP2: Handle oversized credential IDs
The code assumes that all credential IDs are 16 bytes long.
However, we might receive a credential ID that was not generated by the
MiniBLE and hence might be greater than 16 bytes. Handle this by
truncating the credential to 16 bytes and let the rest of the code
handle credential not found and return error.