Add SMC decode README and OSX details
I bumped into some minor issues while attempting to get the SMC decode
script to work, so I wanted to contribute some additional details to help
others in the form of a README:

* The existing instructions were at the top of SMC decode script, so to make
it easier to find this was moved into the README.
* The existing instructions were Windows specific, but this script works
fine on other systems. I tested it on OSX and added a separate section to
the README. Also renamed the script from smc_decode_win.py to smc_decode.py
* Added a section to README that shows how to connect the Mini BLE card
to the card reader. My method was a bit destructive, but at least works.